+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Fix for as.crf when loaded from file and adding more arguments than just the file
 - added txt_feature as a simple feature extraction to identify if a word is capitalised, an email, an url or a number
+- src/cqdb/src/lookup3.c, fix address sanitizer issue 
 
 # CHANGES IN crfsuite VERSION 0.1.1